Fix bug in Usage instructions
'value' is not defined and the captured camera response resides in the 'image' variable.  Furthermore, looks like at initial load of the component the image can be None until browser camera access is provided and the camera component is fully initialized and fails on streamlit's Image component validating image.format without checking for image being valid.